+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/CirculatingSupply__1.java
@@ -31,7 +31,7 @@ public class CirculatingSupply__1 {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(CirculatingSupply__1 .class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(CirculatingSupply__1 .class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("additionalProperties");
         sb.append('=');
         sb.append(((this.additionalProperties == null)?"<null>":this.additionalProperties));
