+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/ExtraordinaryItems.java
@@ -31,7 +31,7 @@ public class ExtraordinaryItems {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(ExtraordinaryItems.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(ExtraordinaryItems.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("additionalProperties");
         sb.append('=');
         sb.append(((this.additionalProperties == null)?"<null>":this.additionalProperties));
