+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/FinancialsChart.java
@@ -59,7 +59,7 @@ public class FinancialsChart {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(FinancialsChart.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(FinancialsChart.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("yearly");
         sb.append('=');
         sb.append(((this.yearly == null)?"<null>":this.yearly));
