+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/IncomeStatementHistory__2.java
@@ -80,7 +80,7 @@ public class IncomeStatementHistory__2 {
     @JsonProperty("totalOtherIncomeExpenseNet")
     private TotalOtherIncomeExpenseNet__1 totalOtherIncomeExpenseNet;
     @JsonProperty("maxAge")
-    private Integer maxAge;
+    private Long maxAge;
     @JsonProperty("discontinuedOperations")
     private DiscontinuedOperations__1 discontinuedOperations;
     @JsonProperty("netIncomeFromContinuingOps")
@@ -291,12 +291,12 @@ public class IncomeStatementHistory__2 {
     }
 
     @JsonProperty("maxAge")
-    public Integer getMaxAge() {
+    public Long getMaxAge() {
         return maxAge;
     }
 
     @JsonProperty("maxAge")
-    public void setMaxAge(Integer maxAge) {
+    public void setMaxAge(Long maxAge) {
         this.maxAge = maxAge;
     }
 
@@ -343,7 +343,7 @@ public class IncomeStatementHistory__2 {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(IncomeStatementHistory__2 .class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(IncomeStatementHistory__2 .class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("researchDevelopment");
         sb.append('=');
         sb.append(((this.researchDevelopment == null)?"<null>":this.researchDevelopment));
