+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/NonRecurring__1.java
@@ -31,7 +31,7 @@ public class NonRecurring__1 {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(NonRecurring__1 .class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(NonRecurring__1 .class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("additionalProperties");
         sb.append('=');
         sb.append(((this.additionalProperties == null)?"<null>":this.additionalProperties));
