+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/OtherStockholderEquity.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class OtherStockholderEquity {
 
     @JsonProperty("raw")
-    private Integer raw;
+    private Long raw;
     @JsonProperty("fmt")
     private String fmt;
     @JsonProperty("longFmt")
@@ -28,12 +28,12 @@ public class OtherStockholderEquity {
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("raw")
-    public Integer getRaw() {
+    public Long getRaw() {
         return raw;
     }
 
     @JsonProperty("raw")
-    public void setRaw(Integer raw) {
+    public void setRaw(Long raw) {
         this.raw = raw;
     }
 
@@ -70,7 +70,7 @@ public class OtherStockholderEquity {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(OtherStockholderEquity.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(OtherStockholderEquity.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("raw");
         sb.append('=');
         sb.append(((this.raw == null)?"<null>":this.raw));
