+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/Quarterly__1.java
@@ -70,7 +70,7 @@ public class Quarterly__1 {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(Quarterly__1 .class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(Quarterly__1 .class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("date");
         sb.append('=');
         sb.append(((this.date == null)?"<null>":this.date));
