+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/financials/RegularMarketChange.java
@@ -57,7 +57,7 @@ public class RegularMarketChange {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(RegularMarketChange.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(RegularMarketChange.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("raw");
         sb.append('=');
         sb.append(((this.raw == null)?"<null>":this.raw));
